removing pixel saturation for heb

+++ SimGeneral/MixingModule/python/hgcalDigitizer_cfi.py
@@ -64,7 +64,8 @@ hgchebackDigitizer = cms.PSet( accumulatorType   = cms.string("HGCDigiProducer")
                                                    shaperN       = cms.double(1.),
                                                    shaperTau     = cms.double(0.),
                                                    caliceSpecific = cms.PSet( nPEperMIP = cms.double(11.0),
-                                                                              nTotalPE  = cms.double(1156),
+                                                                              #1156 pixels => saturation ~300MIP
+                                                                              nTotalPE  = cms.double(11560),
                                                                               xTalk     = cms.double(0.25),
                                                                               sdPixels  = cms.double(3.0) )
                                                    )
